### Description:
Make a call to synapse in the repository instead of using mocked data.

### PR Checklist:
- [ ] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [ ] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [ ] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [ ] Infrastructure/Configuration

### Additional Details (optional)
_<Add more details as necessary to help reviewers better understand the changes...>_

Related work items: #2836022

+++ src/ApiService/DataTransferObjects/HealthControls/HealthControlProperties.cs
@@ -46,7 +46,7 @@ public class HealthControlProperties
     /// Current score.
     /// </summary>
     [ReadOnly(true)]
-    public int CurrentScore { get; internal set; }
+    public double CurrentScore { get; internal set; }
 
     /// <summary>
     /// Target score.
